backend/requirements.txt
# Essential packages only - Python 3.12 compatible
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlmodel==0.0.14
sqlalchemy==2.0.36
alembic==1.12.1
psycopg[binary]>=3.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Cloudinary
cloudinary==1.36.0

# Utilities
python-dotenv==1.0.0
pydantic>=2.7.0,<3.0.0 
pydantic-settings==2.0.3
gunicorn==21.2.0

# File handling
aiofiles==23.2.1

# Date and time
python-dateutil==2.8.2

# AI and Machine Learning (Groq API) - Updated to stable version
groq>=0.3.0,<0.4.0
# Remove conflicting langchain packages and use direct Groq API instead
# langchain==0.0.350  # REMOVED - causing conflicts
# langchain-community==0.0.20  # REMOVED - causing conflicts
# langchain-core==0.1.23  # REMOVED - causing conflicts
# langchain-groq==0.1.1  # REMOVED - causing conflicts
#chromadb==0.4.18
#sentence-transformers==2.2.2
#pinecone-client==2.2.4
#transformers==4.56.0
#torch==2.8.0
#numpy==1.26.4